agentic-tools-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-tools-core"
version = "0.1.0"
description = "Shared contracts and runtime for local agentic recruiting tools"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[tool.setuptools.packages.find]
include = ["agentic_tools_core*"]
